refactor to seperate the layout from the editor functionality
